+++ include/ambition/7kaaInterface/input.hh
@@ -41,6 +41,13 @@ enum class Action {
   Common_Confirm,
   Common_Reward,
 
+  Speed_Decrease_Small,
+  Speed_Decrease_Standard,
+  Speed_Increase_Small,
+  Speed_Increase_Standard,
+  Speed_SetToUnlimited,
+  Speed_TogglePause,
+
   Spy_Assassinate,
   Spy_Bribe,
   Spy_Capture,
@@ -157,6 +164,10 @@ void detectScenarioScroll(
   SlideVBar& scrollBar,
   SlideVBar& descriptionScrollBar,
   int& refreshFlag
+);
+
+bool detectSpeedChangeKeys(
+  const unsigned int keyCode
 );
 
 bool detectSpyScroll(