+++ src/ambition/version_details.cc
@@ -19,6 +19,14 @@ std::vector<VersionUpdates> getVersionDetails(
     .identifier = IDENTIFIER_PREVIEW,
     .sections = {
       {
+        .heading = HEADING_ADDED,
+        .items = {
+          _("Added access to speeds 9 and 10 using the `9` and `0` keys.  Unlimited speed has been moved to `\\`, and pause toggle is already done with Space.  These can be rebound using the `KEYEVENT_ _SPEED_ _SET_ _TO_ _UNLIMITED` and `KEYEVENT_ _SPEED_ _TOGGLE_ _PAUSE` bindings in the config.txt file."),
+          _("Added access to cycling through all speeds 1–32 using the `-` and `=` keys.  These can be rebound using the `KEYEVENT_ _SPEED_ _DECREASE_ _STANDARD` and `KEYEVENT_ _SPEED_ _INCREASE_ _STANDARD` bindings in the config.txt file."),
+          _("Added access to cycling through all the intermediate speeds (there are three steps from each standard speed level to the next; and the two intermediate speeds are shown as .3 and .7).  These can be accessed using the `shift+-` and `shift+=` keys and can be rebound using the `KEYEVENT_ _SPEED_ _DECREASE_ _SMALL` and `KEYEVENT_ _SPEED_ _INCREASE_ _SMALL` bindings in the config.txt file."),
+        },
+      },
+      {
         .heading = HEADING_CHANGED,
         .items = {
           _("Select sortied units only if the Fort is still selected, instead of overriding the selection, preventing accidentally giving orders to the wrong units in multiplayer."),