+++ CHANGELOG.md
@@ -10,6 +10,22 @@ speaking, considered to be the player's experience.
 
 ## [Unreleased]
 
+### Added
+
+- Added access to speeds 9 and 10 using the `9` and `0` keys.  Unlimited speed
+  has been moved to `\`, and pause toggle is already done with Space.  These can
+  be rebound using the `KEYEVENT_SPEED_SET_TO_UNLIMITED` and
+  `KEYEVENT_SPEED_TOGGLE_PAUSE` bindings in the config.txt file.
+- Added access to cycling through all speeds 1–32 using the `-` and `=` keys.
+  These can be rebound using the `KEYEVENT_SPEED_DECREASE_STANDARD` and
+  `KEYEVENT_SPEED_INCREASE_STANDARD` bindings in the config.txt file.
+- Added access to cycling through all the intermediate speeds (there are three
+  steps from each standard speed level to the next; and the two intermediate
+  speeds are shown as .3 and .7).  These can be accessed using the `shift+-` and
+  `shift+=` keys and can be rebound using the
+  `KEYEVENT_SPEED_DECREASE_SMALL` and `KEYEVENT_SPEED_INCREASE_SMALL` bindings
+  in the config.txt file.
+
 ### Changed
 
 - Select sortied units only if the Fort is still selected, instead of overriding